on teacher launch, don't load default unit
this also includes various refactoring to simplify things
