+++ src/index.tsx
@@ -6,7 +6,6 @@ import { AppComponent } from "./components/app";
 import { getAppMode } from "./lib/auth";
 import { urlParams } from "./utilities/url-params";
 import { QAClear } from "./components/qa-clear";
-import { setPageTitle } from "./lib/misc";
 import { getBearerToken, initializeAuthorization } from "./utilities/auth-utils";
 import { removeLoadingMessage, showLoadingMessage } from "./utilities/loading-utils";
 
@@ -25,7 +24,6 @@ if (!redirectingToAuthDomain) {
     );
   } else {
     initializeApp(appMode).then((stores) => {
-      stores.unitLoadedPromise.then(() => setPageTitle(stores));
       stores.ui.setShowDemoCreator(!!stores.showDemoCreator);
 
       ReactDOM.render(
