+++ src/lib/db.ts
@@ -153,7 +153,7 @@ export class DB {
           this.firebase.setFirebaseUser(firebaseUser);
           this.firestore.setFirebaseUser(firebaseUser);
           if (!options.dontStartListeners) {
-            const { persistentUI, user, db, unitLoadedPromise} = this.stores;
+            const { persistentUI, user, db, problemLoadedPromise } = this.stores;
             // Start fetching the persistent UI. We want this to happen as early as possible.
             persistentUI.initializePersistentUISync(user, db);
 
@@ -161,7 +161,7 @@ export class DB {
             // Before they can be started  we need to wait for the unit to be loaded,
             // since it includes the list of tile types being registered.
             // We need those types to be registered so the listeners can safely create documents.
-            unitLoadedPromise.then(() => {
+            problemLoadedPromise.then(() => {
               this.listeners.start().then(resolve).catch(reject);
             });
           }
