+++ src/models/stores/stores.ts
@@ -26,6 +26,7 @@ import { NavTabModelType } from "../view/nav-tabs";
 import { Bookmarks } from "./bookmarks";
 import { SortedDocuments } from "./sorted-documents";
 import { removeLoadingMessage, showLoadingMessage } from "../../utilities/loading-utils";
+import { problemLoaded } from "../../lib/misc";
 
 export interface IStores extends IBaseStores {
   problemPath: string;
@@ -36,9 +37,9 @@ export interface IStores extends IBaseStores {
   studentWorkTabSelectedGroupId: string | undefined;
   setAppMode: (appMode: AppMode) => void;
   initializeStudentWorkTab: () => void;
-  setUnitAndProblem: (unitId: string | undefined, problemOrdinal?: string) => Promise<void>;
+  loadUnitAndProblem: (unitId: string | undefined, problemOrdinal?: string) => Promise<void>;
   sortedDocuments: SortedDocuments;
-  unitLoadedPromise: Promise<void>;
+  problemLoadedPromise: Promise<void>;
 }
 
 export interface ICreateStores extends Partial<IStores> {
@@ -78,7 +79,7 @@ class Stores implements IStores{
   serialDevice: SerialDevice;
   userContextProvider: UserContextProvider;
   sortedDocuments: SortedDocuments;
-  unitLoadedPromise: Promise<void>;
+  problemLoadedPromise: Promise<void>;
 
   constructor(params?: ICreateStores){
     // This will mark all properties as observable
@@ -92,10 +93,18 @@ class Stores implements IStores{
     this.isPreviewing = params?.isPreviewing || false;
     this.appVersion = params?.appVersion || "unknown";
     this.appConfig = params?.appConfig || AppConfigModel.create();
-    // for testing, we create a null problem or investigation if none is provided
+
+    // To keep the code simple, we create a null unit, investigation, and problem if they aren't provided.
+    // Code that needs the real unit, investigation, or problem should wait on
+    // the `problemLoadedPromise`.
+    this.unit = params?.unit || UnitModel.create({code: "NULL", title: "Null Unit"});
     this.investigation = params?.investigation ||
       InvestigationModel.create({ ordinal: 0, title: "Null Investigation" });
-    this.problem = params?.problem || ProblemModel.create({ ordinal: 0, title: "Null Problem" });
+    const defaultProblem = ProblemModel.create({ ordinal: 0, title: "Null Problem" });
+    // CHECKME: do we ever pass the problem?
+    // If we do, then the problemLoadedPromise will resolve immediately
+    this.problem = params?.problem || defaultProblem;
+
     this.user = params?.user || UserModel.create({ id: "0" });
     this.groups = params?.groups || GroupsModel.create({ acceptUnknownStudents: params?.isPreviewing });
     this.groups.setEnvironment(this);
@@ -103,7 +112,6 @@ class Stores implements IStores{
     this.db = params?.db || new DB();
     this.documents = params?.documents || createDocumentsModelWithRequiredDocuments(requiredDocumentTypes);
     this.networkDocuments = params?.networkDocuments || DocumentsModel.create({});
-    this.unit = params?.unit || UnitModel.create({code: "NULL", title: "Null Unit"});
     const demoName = params?.demoName || this.appConfig.appName;
     this.demo = params?.demo || DemoModel.create({name: demoName, class: {id: "0", name: "Null Class"}});
     this.showDemoCreator = params?.showDemoCreator || false;
@@ -127,6 +135,12 @@ class Stores implements IStores{
     this.userContextProvider = new UserContextProvider(this);
     this.bookmarks = new Bookmarks({db: this.db});
     this.sortedDocuments = new SortedDocuments(this);
+
+    // It would be cleaner if we could just leave the problem undefined
+    // until it was loaded for real. However there are lots of components that expect the
+    // problem to be defined. By using a default problem these components can render with
+    // the default problem and then update when it is set for real.
+    this.problemLoadedPromise = when(() => this.problem !== defaultProblem);
   }
 
   get tabsToDisplay() {
@@ -192,7 +206,7 @@ class Stores implements IStores{
   // However typing the yield statements is difficult. Also flows
   // in MobX are slightly different than flows in MST, so there might
   // be some weird interactions with action tracking if we mix them.
-  async setUnitAndProblem(unitId: string | undefined, problemOrdinal?: string) {
+  async loadUnitAndProblem(unitId: string | undefined, problemOrdinal?: string) {
     const { appConfig, persistentUI } = this;
     showLoadingMessage("Loading curriculum content");
     let unitJson = await getUnitJson(unitId, appConfig);
@@ -241,6 +255,9 @@ class Stores implements IStores{
         this.investigation = investigation;
         this.problem = problem;
       }
+
+      problemLoaded(this, _problemOrdinal);
+
       persistentUI.setProblemPath(this.problemPath);
 
       // Set the active tab to be the first tab (unless active tab is already set by persistent UI)
